backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.49
pyodbc==5.3.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
